requirements.txt
boto==2.49.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dj-database-url==0.5.0
enum34==1.1.6
Flask==1.0.2
Flask-Assets==0.12
flask-mongoengine==0.9.5
Flask-Scss==0.5
Flask-Session==0.3.1
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mongoengine==0.16.3
nltk==3.4
oauthlib==3.0.1
pathlib==1.0.1
Pillow==5.4.1
psycopg2==2.7.7
pymongo==3.7.2
pyScss==1.3.5
python-dotenv==0.10.1
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
singledispatch==3.4.0.3
six==1.12.0
urllib3==1.24.1
virtualenv==16.3.0
webassets==0.12.1
Werkzeug==0.14.1
whitenoise==4.1.2
WTForms==2.2.1